fix: Environment vars: allow them to work (#184)
* Env vars: don't overrwite them if there's no config

Previously, commands like `packer deploy` would *require* there to be a
packer.env.json config file or it'd crash. It seems like env vars were
supposed to be allowed as a way to configure packer. But, if you *did*
provide a config file, all your env vars would be overridden.

Here we print an error message if the packer.env.json file can't be
found, but continue, assuming that the environment contains the needed
vars.

* env: fix error messages

These were using a non-existent (maybe old?) config value. The correct
one is env.keys.id (holding the name of the ENV var for the theme id).

Accessing this incorrect value caused *another* "value not found" error.